Data import functions

Some functions to import data from held in on-line data archives.

Usage

importUKNAN(network_id, uka_id, pollutant_id, year, ..., silent = TRUE)

importUKNAMetals(uka_id, pollutant_id, year, ..., silent = TRUE)

importUKNAPAHs(uka_id, pollutant_id, year, ..., silent = TRUE)

importUKNAHCs(uka_id, pollutant_id, year, ..., silent = TRUE)

Arguments

network_id

(importUKNAN only; character) The identifier code for the network archive to request data from.

uka_id

(character) The 'UKA' identifier code(s) for the site(s) the data is to be downloaded from.

pollutant_id

(numeric) The pollutant identifier(s) of the pollutant(s) to download data for.

year

(numeric) The year(s) to download data for.

...

(other arguments) currently ignored.

silent

(logical) Return data silently? If TRUE (default) a NULL is returned without warning if the requested data is not available; if FALSE, options are suggested.

Value

The import functions typically return the request data as a data.frame if data is available for requested site/pollutant/year combination. In some cases the data.frame may be set up for use with use with other packages.

importUKNAN: Generic import function for data from DEFRA's Non-Automatic Networks. (In-development and largely untested; currently recommend using dedicated import if below or getting in contact.)

See https://uk-air.defra.gov.uk/networks/network-info?view=non-automatic.

importUKNAMetals: import function for metal data from the UK Non-automatic Heavy Metals Network. Previously called the Urban/Industrial Heavy Metals Network and Rural Heavy Metals Network.

See https://uk-air.defra.gov.uk/networks/network-info?view=metals for details and source information.

importUKNAPAHs: import function for metal data from the UK Non-automatic Polycyclic Aromatic Hydrocarbons (PAH) Network.

see https://uk-air.defra.gov.uk/networks/network-info?view=pah for details and source information.

importUKNAHCs: import function for metal data from the UK Non-automatic Hydrocarbon Network.

see https://uk-air.defra.gov.uk/networks/network-info?view=nahc for details and source information.

Note

These functions were developed while working on projects that required multiple downloads of data from these sources. The code is shared freely but with the caveat that I am not the data provider. So:

(a) You should cite both this package and the data source (identified for each function), and

(b) You should be aware that they only work for as long as the structure of the data in the archive remains as it was when the code was written, and neither I or you have any control in that...