Local Time Offsets

`ref.hour.offset` is a lookup table of local time offsets.

Usage

ref.hour.offset

Format

A `data.frame` 38 locations and associated time offsets:

local

(Character) The location for which an offset value is reported.

hour.offset

(Numeric) The time offset in hours associated with local.

Source

NOAA assignment table: http://www.srrb.noaa.gov/highlights/sunrise/timezone.html.

Details

ref.hour.offset is a lookup table for use with calcNOAASunrise and hour.offset.

Examples

#basic structure
head(ref.hour.offset)
#>               local hour.offset
#> 1 GREENWICH ENGLAND         0.0
#> 2      SE GREENLAND         1.0
#> 3    ATLANTIC OCEAN         2.0
#> 4    GREENLAND TIME         3.0
#> 5 NEWFOUNDLAND TIME         3.5
#> 6     ATLANTIC TIME         4.0
#usage
hour.offset("greenwich")
#> [1] 0
hour.offset("india")
#> [1] -5.5
hour.offset("hawaii")
#> [1] 10