packaging.tools

Code written when making packages.

splat_pack(pkg = ".", overwrite = FALSE, pkg.dir = ".", ...)

splat_description(
  title = ".",
  description = ".",
  pkg = ".",
  overwrite = FALSE,
  ...
)

splat_citation(overwrite = FALSE, pkg = ".", ...)

splat_package_intro(
  name = ".",
  pkg = ".",
  overwrite = FALSE,
  title = ".",
  description = ".",
  ...
)

splat_package_readme(
  name = ".",
  pkg = ".",
  overwrite = FALSE,
  title = ".",
  description = ".",
  ...
)

splat_package_news(
  name = ".",
  pkg = ".",
  overwrite = FALSE,
  title = ".",
  description = ".",
  ...
)

splat_function(
  fun = ".",
  overwrite = FALSE,
  pkg = ".",
  title = ".",
  name = ".",
  ...
)

splat_data(data = ".", overwrite = FALSE, pkg = ".", ...)

splat_date(title = ".", date = ".", pkg = ".", overwrite = FALSE, ...)

splat_version(
  title = ".",
  version = ".",
  increase = ".",
  pkg = ".",
  overwrite = FALSE,
  ...
)

Arguments

pkg

(character) package name.

overwrite

(logical) Overwrite file if already present, default typically FALSE.

pkg.dir

(character) main packaging directory.

...

Other arguments passed on.

title

(character) title of package or document template built as part of splat.

description

(character) text for package description.

name

(character) name of document template if different from item (e.g. function) being added to package.

fun

(character) name of function to add to package.

data

(character) name of data to add to package.

date

(character) date in 'YYYY-MM_DD' format.

version

(character) version in 'n.n.n' format.

increase

(character or numeric) version increase, either '1', '0.1' or '0.0.1' or 1, 2, or 3 from first, second or third level version increase.